Revert "Revert "Accounts Module user changes as with the other modules""
This reverts commit 91653beffd091e337cc120237c50e9bd93b3a799.

+++ freightbit/src/main/webapp/core/group/groupListView.jsp
@@ -17,7 +17,14 @@
         <div class="panel panel-primary">
 
             <div class="panel-heading">
-                <h3 class="panel-title"><i class="fa fa-list"></i> Group List</h3>
+                <h3 class="panel-title" style="float:left;top: 10px;"><i class="fa fa-list"></i> Group List
+                </h3>
+                <span class="pull-right">
+                    <button type="button" class="btn btn-primary new-booking"
+                            onclick="location.href='loadAddGroupPage'">
+                        <i class="fa fa-male"></i> New Group
+                    </button>
+                </span>
             </div>
 
             <div class="panel-body">
@@ -37,17 +44,13 @@
                                         <s:url var="loadEditGroupUrl" action="loadEditGroupPage">
                                             <s:param name="groupCodeParam" value="#attr.group.groupCode"></s:param>
                                         </s:url>
-                                        <s:a href="%{loadEditGroupUrl}" class="icon-action-link" rel="tooltip"
-                                             title="Edit this group"><img src="../includes/images/edit-user.png"
-                                                                          class="icon-action circ-icon"> </s:a>
+                                        <s:a href="%{loadEditGroupUrl}" class="icon-action-link" rel="tooltip" title="Edit this group">
+                                        <!-- <img src="../includes/images/edit-user.png" class="icon-action circ-icon"> </s:a> --><i class="fa fa-pencil"></i>
                                         <s:url var="deleteGroupUrl" action="deleteGroup">
                                             <s:param name="groupCodeParam" value="#attr.group.groupCode"></s:param>
                                         </s:url>
-                                        <s:a class="icon-action-link" href="%{deleteGroupUrl}" rel="tooltip"
-                                             title="Delete this group"
-                                             onclick="return confirm('Do you really want to delete?');"><img
-                                                src="../includes/images/remove-user.png"
-                                                class="icon-action circ-icon"> </s:a>
+                                        <s:a class="icon-action-link" href="%{deleteGroupUrl}" rel="tooltip" title="Delete this group" onclick="return confirm('Do you really want to delete?');">
+                                        <!-- <img src="../includes/images/remove-user.png" class="icon-action circ-icon"> </s:a>--><i class="fa fa-trash-o"></i>
                                     </display:column>
                                 </td>
                             </display:table>
@@ -60,8 +63,10 @@
             <div class="panel-footer">
 
                 <span class="pull-right">
-                <a href="<s:url action='loadAddGroupPage'/>" class="icon-action-link"><img
-                        src="../includes/images/add-user.png" class="icon-action circ-icon"> </a>
+                    <button type="button" class="btn btn-primary new-booking"
+                            onclick="location.href='loadAddGroupPage'">
+                        <i class="fa fa-male"></i> New Group
+                    </button>
                 </span>
 
             </div>
